--- back/students-ms/src/main/java/com/iamind/students_ms/StudentsMsApplication.java
@@ -1,13 +1,0 @@
-package com.iamind.students_ms;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StudentsMsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(StudentsMsApplication.class, args);
-    }
-
-}
