--- back/students-ms/src/main/java/com/iamind/students_ms/controller/StudentController.java
@@ -1,45 +1,0 @@
-package com.iamind.students_ms.controller;
-
-import com.iamind.students_ms.dto.StudentRequestDTO;
-import com.iamind.students_ms.dto.StudentResponseDTO;
-import com.iamind.students_ms.service.StudentService;
-import jakarta.validation.Valid;
-import lombok.AllArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@AllArgsConstructor
-@RestController
-@RequestMapping("/api/students")
-public class StudentController {
-
-    private final StudentService studentService;
-
-    @GetMapping
-    public List<StudentResponseDTO> getAllStudents() {
-        return studentService.getAllStudents();
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<StudentResponseDTO> getStudentById(@PathVariable Long id) {
-        return ResponseEntity.ok(studentService.getStudentById(id));
-    }
-
-    @PostMapping
-    public ResponseEntity<StudentResponseDTO> createStudent(@Valid @RequestBody StudentRequestDTO dto) {
-        return ResponseEntity.ok(studentService.createStudent(dto));
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<StudentResponseDTO> updateStudent(@PathVariable Long id, @Valid @RequestBody StudentRequestDTO dto) {
-        return ResponseEntity.ok(studentService.updateStudent(id, dto));
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteStudent(@PathVariable Long id) {
-        studentService.deleteStudent(id);
-        return ResponseEntity.noContent().build();
-    }
-}
