--- back/students-ms/src/main/java/com/iamind/students_ms/dto/StudentResponseDTO.java
@@ -1,14 +1,0 @@
-package com.iamind.students_ms.dto;
-
-import java.time.LocalDate;
-
-public record StudentResponseDTO(
-        Long id,
-        String firstName,
-        String lastName,
-        LocalDate dateOfBirth,
-        String school,
-        String registrationNumber,
-        String email
-) {
-}
