--- back/students-ms/src/main/java/com/iamind/students_ms/exception/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package com.iamind.students_ms.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
-
