--- back/students-ms/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-spring.application.name=students-ms
-# Temporary H2 DB config
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.hibernate.ddl-auto=update
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
