--- back/students-ms/src/test/java/com/iamind/students_ms/StudentsMsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.iamind.students_ms;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class StudentsMsApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
