+++ back/user-ms/src/main/java/com/iamind/user_ms/UserMsApplication.java
@@ -1,0 +1,13 @@
+package com.iamind.user_ms;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserMsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UserMsApplication.class, args);
+    }
+
+}
