+++ back/user-ms/src/main/java/com/iamind/user_ms/dto/UserResponseDTO.java
@@ -1,0 +1,14 @@
+package com.iamind.user_ms.dto;
+
+import java.time.LocalDate;
+
+public record UserResponseDTO(
+        Long id,
+        String firstName,
+        String lastName,
+        LocalDate dateOfBirth,
+        String school,
+        String registrationNumber,
+        String email
+) {
+}
