+++ back/user-ms/src/main/java/com/iamind/user_ms/exception/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.iamind.user_ms.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
+
