+++ back/user-ms/src/main/java/com/iamind/user_ms/model/User.java
@@ -1,0 +1,40 @@
+package com.iamind.user_ms.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Entity
+@Table(name = "students")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String firstName;
+
+    @Column(nullable = false)
+    private String lastName;
+
+    @Column(nullable = false)
+    private LocalDate dateOfBirth;
+
+    @Column(nullable = false)
+    private String school;
+
+    @Column(nullable = false, unique = true)
+    private String registrationNumber; // RA
+
+    @Column(nullable = false, unique = true)
+    private String email;
+}
