+++ back/user-ms/src/main/java/com/iamind/user_ms/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.iamind.user_ms.repository;
+
+import com.iamind.user_ms.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}
+
