+++ back/user-ms/src/main/java/com/iamind/user_ms/service/UserService.java
@@ -1,0 +1,18 @@
+package com.iamind.user_ms.service;
+
+import com.iamind.user_ms.dto.UserRequestDTO;
+import com.iamind.user_ms.dto.UserResponseDTO;
+
+import java.util.List;
+
+public interface UserService {
+    List<UserResponseDTO> getAllUsers();
+
+    UserResponseDTO getUserById(Long id);
+
+    UserResponseDTO createUser(UserRequestDTO userRequestDTO);
+
+    UserResponseDTO updateUser(Long id, UserRequestDTO userRequestDTO);
+
+    void deleteUser(Long id);
+}
