+++ back/user-ms/src/test/java/com/iamind/user_ms/UserMsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.iamind.user_ms;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserMsApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
